[BACKLOG-197] - Document Rest Endpoints for User Roles. Refactored SystemResourceUtil to be named SystemResourceService and moved the class to the service folder. This change is to match the naming conventions of the other classes.
